修正: 疎結合化
