+++ backend/src/qni/qiskit_runner.py
@@ -1,16 +1,17 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable, TypedDict, cast
+from typing import TYPE_CHECKING, Callable, TypedDict
 
 import numpy as np
 
 if TYPE_CHECKING:
     from qiskit.result import Result  # type: ignore
 
-from qiskit import ClassicalRegister, QuantumCircuit, transpile  # type: ignore
-from qiskit.circuit.library import HGate, SdgGate, SGate, TdgGate, TGate, XGate, YGate, ZGate  # type: ignore
+from qiskit import QuantumCircuit, transpile  # type: ignore
+from qiskit.qasm3 import dumps  # type: ignore
 from qiskit_aer import AerSimulator  # type: ignore
 
+from qni.qiskit_circuit_builder import QiskitCircuitBuilder
 from qni.types import (
     DeviceType,
     MeasuredBits,
@@ -34,7 +35,6 @@ OperationMethod = Callable[[QuantumCircuit, BasicOperation | ControllableOperati
 
 
 class QiskitRunner:
-    _PAIR_OPERATION_COUNT = 2
     _STATEVECTOR_LABEL = "state_at_until_step"
 
     def __init__(self, logger=None):
@@ -115,18 +115,39 @@ class QiskitRunner:
 
     def _process_step_operations(self, qubit_count: int, until_step_index: int) -> QuantumCircuit:
         circuit = QuantumCircuit(qubit_count)
+        circuit_builder = QiskitCircuitBuilder()
 
         for step_index, step in enumerate(self.steps):
             if len(step) == 0:
                 circuit.id(list(range(qubit_count)))
 
             for operation in step:
-                self._apply_operation(circuit, operation)
+                circuit_builder.apply_operation(circuit, operation)
 
             if step_index == until_step_index:
                 circuit.save_statevector(label=self._STATEVECTOR_LABEL)
 
         return circuit
+
+    def build_circuit_for_export(
+        self,
+        steps: list,
+        qubit_count: int,
+    ) -> QuantumCircuit:
+        circuit = QuantumCircuit(qubit_count)
+        circuit_builder = QiskitCircuitBuilder()
+
+        for step in steps:
+            if len(step) == 0:
+                circuit.id(list(range(qubit_count)))
+
+            for operation in step:
+                circuit_builder.apply_operation(circuit, operation)
+
+        return circuit
+
+    def convert_to_qasm3(self, circuit: QuantumCircuit) -> str:
+        return dumps(circuit)
 
     class UnknownOperationError(ValueError):
         """
@@ -138,124 +159,6 @@ class QiskitRunner:
 
         def __init__(self, operation_type):
             super().__init__(f"Unknown operation: {operation_type}")
-
-    def _apply_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        operation_type = operation["type"]
-        operation_methods: dict[str, OperationMethod] = {
-            "H": self._apply_h_operation,
-            "X": self._apply_x_operation,
-            "Y": self._apply_y_operation,
-            "Z": self._apply_z_operation,
-            "X^½": self._apply_rnot_operation,
-            "S": self._apply_s_operation,
-            "S†": self._apply_s_dagger_operation,
-            "T": self._apply_t_operation,
-            "T†": self._apply_t_dagger_operation,
-            "Swap": self._apply_swap_operation,
-            "•": self._apply_controlled_z_operation,
-            "|0>": self._apply_write0,
-            "|1>": self._apply_write1,
-            "Measure": self._apply_measure_operation,
-        }
-
-        if operation_type in operation_methods:
-            operation_methods[operation_type](circuit, operation)
-        else:
-            raise self.UnknownOperationError(operation_type)
-
-    def _apply_h_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, HGate())
-        else:
-            circuit.h(operation["targets"])
-
-    def _apply_x_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            operation = cast(ControllableOperation, operation)
-            for target in operation["targets"]:
-                circuit.mcx(operation["controls"], target)
-        else:
-            circuit.x(operation["targets"])
-
-    def _apply_y_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, YGate())
-        else:
-            circuit.y(operation["targets"])
-
-    def _apply_z_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, ZGate())
-        else:
-            circuit.z(operation["targets"])
-
-    def _apply_rnot_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, XGate().power(1 / 2))
-        else:
-            circuit.append(XGate().power(1 / 2), qargs=operation["targets"])
-
-    def _apply_s_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, SGate())
-        else:
-            circuit.s(operation["targets"])
-
-    def _apply_s_dagger_operation(
-        self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation
-    ) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, SdgGate())
-        else:
-            circuit.sdg(operation["targets"])
-
-    def _apply_t_operation(self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, TGate())
-        else:
-            circuit.t(operation["targets"])
-
-    def _apply_t_dagger_operation(
-        self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation
-    ) -> None:
-        if "controls" in operation:
-            self._apply_controlled_u(circuit, operation, TdgGate())
-        else:
-            circuit.tdg(operation["targets"])
-
-    def _apply_swap_operation(self, circuit: QuantumCircuit, operation: BasicOperation) -> None:
-        if len(operation["targets"]) == self._PAIR_OPERATION_COUNT:
-            circuit.swap(operation["targets"][0], operation["targets"][1])
-        else:
-            circuit.id(operation["targets"])
-
-    def _apply_controlled_z_operation(self, circuit: QuantumCircuit, operation: BasicOperation) -> None:
-        if len(operation["targets"]) >= self._PAIR_OPERATION_COUNT:
-            u = ZGate().control(num_ctrl_qubits=len(operation["targets"]) - 1)
-            circuit.append(u, qargs=operation["targets"])
-        else:
-            circuit.id(operation["targets"])
-
-    def _apply_write0(self, circuit: QuantumCircuit, operation: BasicOperation) -> None:
-        circuit.reset(operation["targets"])
-
-    def _apply_write1(self, circuit: QuantumCircuit, operation: BasicOperation) -> None:
-        circuit.reset(operation["targets"])
-        circuit.x(operation["targets"])
-
-    def _apply_measure_operation(self, circuit: QuantumCircuit, operation: BasicOperation) -> None:
-        creg = ClassicalRegister(circuit.num_qubits)
-        circuit.add_register(creg)
-        for target in operation["targets"]:
-            circuit.measure(target, creg[target])
-
-    def _apply_controlled_u(
-        self, circuit: QuantumCircuit, operation: BasicOperation | ControllableOperation, gate
-    ) -> None:
-        operation = cast(ControllableOperation, operation)
-        u = gate.control(num_ctrl_qubits=len(operation["controls"]))
-        for target in operation["targets"]:
-            circuit.append(u, qargs=operation["controls"] + [target])
 
     def _run_backend(self, device: DeviceType) -> Result:
         backend = AerSimulator(method="statevector")
